+++ src/App.js
@@ -1,23 +1,27 @@
 import { BrowserRouter as Router, Route } from 'react-router-dom';
+import { withRouter } from 'react-router';
 import Navbar from './components/Navbar';
+import Home from './components/Home';
 import NewPage from './components/NewPage';
 import Footer from './components/Footer';
 
 export default function App() {
   return (
     <div className="dark:bg-gray-900 dark:text-white">
+      <Navbar />
       <Router>
-        <Navbar />
         <div className="p-2">
           <Route exact path="/">
-            <h1>Main Page</h1>
+            <Home />
           </Route>
           <Route path="/new/:path">
             <NewPage />
           </Route>
         </div>
-        <Footer />
       </Router>
+      <Footer />
     </div>
   );
 }
+
+withRouter(Navbar);